+++ core/src/no/ntnu/game/Views/MainMenuScreen.java
@@ -91,8 +91,8 @@ public class MainMenuScreen extends Screen {
     public void render(SpriteBatch sb) {
         final String HIGHSCORE_STRING = "Your highscore: " + PLAYER_CONTROLLER.getPlayer().getHighScore().toString();
         final String FASTEST_STRING = "Your fastest time: "
-                + (Float.isInfinite(PLAYER_CONTROLLER.getPlayer().getFastestTime()) ? "0"
-                        : PLAYER_CONTROLLER.getPlayer().getFastestTime().toString());
+        + (Float.isInfinite(PLAYER_CONTROLLER.getPlayer().getFastestTime()) ? "0"
+                : Integer.toString(PLAYER_CONTROLLER.getPlayer().getFastestTime().intValue()));
         final String WELCOME_STRING = "Welcome " + PLAYER_CONTROLLER.getPlayer().getUsername() + "!";
         final float CENTER_WELCOME_X = calculateCenterX(WELCOME_STRING, font);
         final float CENTER_USER_HIGHSCORE_X = calculateCenterX(HIGHSCORE_STRING, font);
