+++ core/src/no/ntnu/game/factory/PowerUp/AbstractPowerUpFactory.java
@@ -1,5 +1,7 @@
 package no.ntnu.game.Factory.PowerUp;
 
+import no.ntnu.game.Models.PowerUp;
+
 public interface AbstractPowerUpFactory {
     PowerUp createPowerUp();
 }
